Rewrite main SPNS application to be a standalone binary
It was rather awkward to have to load the main SPNS via Python; this
rewrites it to create a first-class `SPNS` binary instead, using an
embedded Python interpreter for the one small part (initial config
parsing) that still relies on Python (for parsing and sharing config
with notifiers).

Also:
- removes the old firebase and older fcm notifier (firebase notification
  is now C++ and lives in its own repo)
- install (new) binary, library, and pybind module, rather than dumping
  them in the top-level project.  The binary + library are installed
  under CMAKE_INSTALL_PREFIX, and the python3 executable goes into the
  python user site library location
  (`~/.local/lib/python3.x/site-packages/spns/core.cpython-....so`)
