+++ spns/notifiers/huawei.py
@@ -2,7 +2,7 @@
 
 from .. import config
 from ..config import logger
-from ..core import SUBSCRIBE
+from spns_hivemind import SUBSCRIBE
 from .util import encrypt_notify_payload, derive_notifier_key, warn_on_except, NotifyStats
 
 from hms.src import push_admin
