+++ spns/web.py
@@ -3,11 +3,12 @@ from . import config
 import coloredlogs
 import uwsgi
 import oxenmq
+from spns_hivemind import core_logger
 from uwsgidecorators import postfork
 
 app = flask.Flask(__name__)
 coloredlogs.install(
-    milliseconds=True, isatty=True, logger=app.logger, level=config.core_logger.get_level()
+    milliseconds=True, isatty=True, logger=app.logger, level=core_logger.get_level()
 )
 
 # Monkey-patch app.get/post/etc. for Flask <2 compatibility; this has to be before the imports,
