+++ systemd/spns-hivemind.service
@@ -11,8 +11,8 @@ WatchdogSec=1min
 WorkingDirectory=/home/push/session-push-notification-server
 LimitNOFILE=16384
 Restart=always
-RestartSec=5s
-ExecStart=/usr/bin/python3 -mspns.hivemind
+RestartSec=2s
+ExecStart=/home/push/bin/SPNS spns.ini
 
 [Install]
 WantedBy=multi-user.target
